requirements.txt
langchain
langchain-community
langchain-core
langchain-text-splitters
faiss-cpu
sentence-transformers
flask
langchain-huggingface
requests
beautifulsoup4